Change kge model API and loading of triples and mappings
The API of the kge models was changed so that only relevant parameters for
the model have to be given at initiation. Parameters that concern the
data, e.g. number of entities and relations, are created within a model
while loading the data and the corresponding embeddings are first
initialized when starting training of the model.
Furthermore, methods for predicting an object given the subject and
relation as well as predicting the subject given an object and relation
were implemented. Two example notebooks using these functionalities are
included.
	modified:   notebooks/training_of_kge_models/NEW_API_Train ConvE on RDF.ipynb
	modified:   notebooks/training_of_kge_models/NEW_API_Train TransE on RDF.ipynb
	modified:   setup.cfg
	modified:   src/pykeen/hpo/random_search.py
	modified:   src/pykeen/kge_models/base.py
	modified:   src/pykeen/kge_models/conv_e.py
	modified:   src/pykeen/kge_models/distmult.py
	modified:   src/pykeen/kge_models/ermlp.py
	modified:   src/pykeen/kge_models/rescal.py
	modified:   src/pykeen/kge_models/structured_embedding.py
	modified:   src/pykeen/kge_models/trans_d.py
	modified:   src/pykeen/kge_models/trans_e.py
	modified:   src/pykeen/kge_models/trans_h.py
	modified:   src/pykeen/kge_models/trans_h_adapted_loss.py
	modified:   src/pykeen/kge_models/trans_r.py
	modified:   src/pykeen/kge_models/unstructured_model.py
	modified:   src/pykeen/utilities/evaluation_utils/metrics_computations.py
	modified:   src/pykeen/utilities/pipeline.py
	modified:   src/pykeen/utilities/triples_creation_utils/instance_creation_utils.py
	modified:   tests/test_models.py